+++ tests/Expectation/ExceptionExpectationTest.php
@@ -4,34 +4,39 @@ declare(strict_types = 1);
 
 namespace hanneskod\readmetester\Expectation;
 
-use hanneskod\readmetester\Runner\Result;
+use hanneskod\readmetester\Runner\OutcomeInterface;
 
 class ExceptionExpectationTest extends \PHPUnit\Framework\TestCase
 {
-    public function testNoException()
+    public function testToString()
     {
-        $expectation = new ExceptionExpectation('foo');
-        $this->assertInstanceOf(
-            Failure::CLASS,
-            $expectation->validate(new Result('', '', null))
-        );
+        $this->assertInternalType('string', (string)new ExceptionExpectation(''));
+    }
+
+    public function testHandles()
+    {
+        $outcome = $this->prophesize(OutcomeInterface::CLASS);
+        $outcome->getType()->willReturn(OutcomeInterface::TYPE_EXCEPTION);
+        $this->assertTrue((new ExceptionExpectation(''))->handles($outcome->reveal()));
     }
 
     public function testWrongException()
     {
-        $expectation = new ExceptionExpectation('foo');
+        $outcome = $this->prophesize(OutcomeInterface::CLASS);
+        $outcome->getPayload()->willReturn(['class' => 'bar']);
         $this->assertInstanceOf(
             Failure::CLASS,
-            $expectation->validate(new Result('', '', new \Exception))
+            (new ExceptionExpectation('foo'))->handle($outcome->reveal())
         );
     }
 
     public function testValidException()
     {
-        $expectation = new ExceptionExpectation('Exception');
+        $outcome = $this->prophesize(OutcomeInterface::CLASS);
+        $outcome->getPayload()->willReturn(['class' => 'foo']);
         $this->assertInstanceOf(
             Success::CLASS,
-            $expectation->validate(new Result('', '', new \Exception))
+            (new ExceptionExpectation('foo'))->handle($outcome->reveal())
         );
     }
 }