+++ tests/Expectation/ReturnTypeExpectationTest.php
@@ -4,50 +4,49 @@ declare(strict_types = 1);
 
 namespace hanneskod\readmetester\Expectation;
 
-use hanneskod\readmetester\Runner\Result;
+use hanneskod\readmetester\Runner\OutcomeInterface;
 
 class ReturnTypeExpectationTest extends \PHPUnit\Framework\TestCase
 {
-    public function testGettypeNoMatch()
+    public function testToString()
     {
-        $expectation = new ReturnTypeExpectation('integer');
+        $this->assertInternalType('string', (string)new ReturnTypeExpectation(''));
+    }
+
+    public function testHandles()
+    {
+        $outcome = $this->prophesize(OutcomeInterface::CLASS);
+        $outcome->getType()->willReturn(OutcomeInterface::TYPE_RETURN);
+        $this->assertTrue((new ReturnTypeExpectation(''))->handles($outcome->reveal()));
+    }
+
+    public function testNoMatch()
+    {
+        $outcome = $this->prophesize(OutcomeInterface::CLASS);
+        $outcome->getPayload()->willReturn(['type' => 'bar', 'class' => 'bar']);
         $this->assertInstanceOf(
             Failure::CLASS,
-            $expectation->validate(new Result('string', ''))
+            (new ReturnTypeExpectation('foo'))->handle($outcome->reveal())
         );
     }
 
-    public function testClassNoMatch()
+    public function testMatchType()
     {
-        $expectation = new ReturnTypeExpectation('integer');
-        $this->assertInstanceOf(
-            Failure::CLASS,
-            $expectation->validate(new Result('string', ''))
-        );
-    }
-
-    public function typeProvider()
-    {
-        return [
-            ['BOOLEAN', true],
-            ['integer', 123],
-            ['double', 1.1],
-            ['string', '123'],
-            ['array', []],
-            ['null', null],
-            ['Exception', new \Exception],
-        ];
-    }
-
-    /**
-     * @dataProvider typeProvider
-     */
-    public function testInteger($expected, $value)
-    {
-        $expectation = new ReturnTypeExpectation($expected);
+        $outcome = $this->prophesize(OutcomeInterface::CLASS);
+        $outcome->getPayload()->willReturn(['type' => 'FOO', 'class' => 'bar']);
         $this->assertInstanceOf(
             Success::CLASS,
-            $expectation->validate(new Result($value, ''))
+            (new ReturnTypeExpectation('foo'))->handle($outcome->reveal())
+        );
+    }
+
+    public function testMatchClass()
+    {
+        $outcome = $this->prophesize(OutcomeInterface::CLASS);
+        $outcome->getPayload()->willReturn(['type' => 'bar', 'class' => 'FOO']);
+        $this->assertInstanceOf(
+            Success::CLASS,
+            (new ReturnTypeExpectation('foo'))->handle($outcome->reveal())
         );
     }
 }