+++ CHANGES.md
@@ -1,3 +1,5 @@
 # Changes in 1.0-beta4
 
 * `@include` annotation replaces `@extends`.
+* Removed `@expectNothing`
+* Now failes if examples produces outcomes that are not handled